Moved common methods to helpers.
