+++ TransXChange.Common/TransXChange.Common.csproj
@@ -12,6 +12,7 @@
     
   <ItemGroup>
     <Folder Include="Extensions\" />
+    <Folder Include="Helpers\" />
     <Folder Include="Models\" />
     <Folder Include="Utils\" />
   </ItemGroup>
