+++ TransXChange.Wales/Program.cs
@@ -3,7 +3,7 @@ using CommandLine.Text;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using TransXChange.Common;
+using TransXChange.Common.Helpers;
 using TransXChange.Common.Models;
 
 namespace TransXChange.Wales
@@ -25,37 +25,37 @@ namespace TransXChange.Wales
             Console.WriteLine(CopyrightInfo.Default);
             Console.WriteLine("");
 
-            Dictionary<string, NAPTANStop> stops = Naptan.Read(options.Naptan);
+            Dictionary<string, NAPTANStop> stops = NaptanHelpers.Read(options.Naptan);
             Console.WriteLine(string.Format("READ: National Public Transport Access Nodes (NaPTAN). Found {0:#,##0.##} stops.", stops.Count));
 
-            Dictionary<string, TXCSchedule> originals = Traveline.ReadWales(stops, options.Traveline, options.Mode, options.Filters, options.Days);
+            Dictionary<string, TXCSchedule> originals = TravelineHelpers.ReadWales(stops, options.Traveline, options.Mode, options.Filters, options.Days);
             Console.WriteLine(string.Format("READ: Traveline National Dataset (TNDS). Found {0:#,##0.##} original schedules.", originals.Count));
 
-            Dictionary<string, TXCSchedule> duplicates = Traveline.ScanDuplicate(originals);
+            Dictionary<string, TXCSchedule> duplicates = TravelineHelpers.ScanDuplicate(originals);
             Console.WriteLine(string.Format("READ: Traveline National Dataset (TNDS). Found {0:#,##0.##} duplicate schedules.", duplicates.Count));
 
             Directory.CreateDirectory(options.Output);
             Console.WriteLine(string.Format("WRITE: {0}", options.Output));
 
-            Gtfs.WriteAgency(originals, duplicates, options.Output);
+            GtfsHelpers.WriteAgency(originals, duplicates, options.Output);
             Console.WriteLine(string.Format("WRITE: {0}", Path.Combine(options.Output, "agency.txt")));
 
-            Gtfs.WriteCalendar(originals, duplicates, options.Output);
+            GtfsHelpers.WriteCalendar(originals, duplicates, options.Output);
             Console.WriteLine(string.Format("WRITE: {0}", Path.Combine(options.Output, "calendar.txt")));
 
-            Gtfs.WriteCalendarDates(originals, duplicates, options.Output);
+            GtfsHelpers.WriteCalendarDates(originals, duplicates, options.Output);
             Console.WriteLine(string.Format("WRITE: {0}", Path.Combine(options.Output, "calendar_dates.txt")));
 
-            Gtfs.WriteRoutes(originals, duplicates, options.Output);
+            GtfsHelpers.WriteRoutes(originals, duplicates, options.Output);
             Console.WriteLine(string.Format("WRITE: {0}", Path.Combine(options.Output, "routes.txt")));
 
-            Gtfs.WriteStops(originals, duplicates, options.Output);
+            GtfsHelpers.WriteStops(originals, duplicates, options.Output);
             Console.WriteLine(string.Format("WRITE: {0}", Path.Combine(options.Output, "stops.txt")));
 
-            Gtfs.WriteStopTimes(originals, duplicates, options.Output);
+            GtfsHelpers.WriteStopTimes(originals, duplicates, options.Output);
             Console.WriteLine(string.Format("WRITE: {0}", Path.Combine(options.Output, "stop_times.txt")));
 
-            Gtfs.WriteTrips(originals, duplicates, options.Output);
+            GtfsHelpers.WriteTrips(originals, duplicates, options.Output);
             Console.WriteLine(string.Format("WRITE: {0}", Path.Combine(options.Output, "trips.txt")));
         }
     }
